requirements.txt
# Deep Learning Framework (CUDA 12.8 optimized for RTX 5090)
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.0+cu128
torchvision==0.22.0+cu128
torchaudio==2.7.0+cu128

# Computer Vision
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
albumentations==1.4.20
pillow==11.0.0

# Object Detection (YOLO)
ultralytics==8.3.50

# Data Processing
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
imbalanced-learn==0.12.4

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Experiment Tracking
tensorboard==2.18.0
wandb==0.18.7

# Configuration
pyyaml==6.0.2
omegaconf==2.3.0
hydra-core==1.3.2

# Utilities
tqdm==4.67.1
colorama==0.4.6
rich==13.9.4

# Model Optimization
timm==1.0.11
efficientnet-pytorch==0.7.1

# Metrics & Evaluation
torchmetrics==1.5.2

# GPU Monitoring
pynvml==11.5.3

# Configuration Management
python-dotenv==1.0.1

# Image Processing
imageio==2.36.1
imageio-ffmpeg==0.5.1

# Multi-processing
joblib==1.4.2

# Logging
loguru==0.7.3
